AnimationDemo: move animation implementation to Brick.Animation
