+++ brick.cabal
@@ -72,6 +72,7 @@ library
   hs-source-dirs:      src
   exposed-modules:
     Brick
+    Brick.Animation
     Brick.AttrMap
     Brick.BChan
     Brick.BorderMap
@@ -128,7 +129,10 @@ library
                        deepseq >= 1.3 && < 1.6,
                        unix-compat,
                        bytestring,
-                       word-wrap >= 0.2
+                       word-wrap >= 0.2,
+                       unordered-containers,
+                       hashable,
+                       time
 
 executable brick-custom-keybinding-demo
   if !flag(demos)
@@ -470,9 +474,6 @@ executable brick-animation-demo
                        microlens-th,
                        microlens-mtl,
                        stm,
-                       unordered-containers,
-                       hashable,
-                       time,
                        mtl
 
 executable brick-custom-event-demo
